rename packge (step -> MonsterRace)
